requirements.txt
Django==1.4.5
MySQL-python==1.2.4
PIL==1.1.7
argparse==1.2.1
distribute==0.6.35
sorl-thumbnail==11.12
wsgiref==0.1.2
